+++ src/.internal/charts/stock/drawing/IconSeries.ts
@@ -135,10 +135,10 @@ export class IconSeries extends PolylineSeries {
 				const dataObject = this.data.getIndex(this.data.length - 1) as any;
 				dataObject.settings = this._getIconTemplate();
 
-				this._increaseIndex();
 				this._di[this._index] = {};
 
 				this._dispatchStockEvent("drawingadded", this._drawingId, this._index);
+				this._increaseIndex();
 			}
 		}
 		this.isDrawing(false);